+++ acf/php/mod-manualinput-settings.php
@@ -47,6 +47,26 @@ if (function_exists('acf_add_local_field_group')) {
             'ui_on_text' => '',
             'ui_off_text' => '',
         ),
+        2 => array(
+            'key' => 'field_68c93c4e001ac',
+            'label' => __('Centrera allt textinnehåll', 'municipio-customisation'),
+            'name' => 'center_content',
+            'aria-label' => '',
+            'type' => 'true_false',
+            'instructions' => __('Centrerar rubrik, brödtext och knapp i korten.', 'municipio-customisation'),
+            'required' => 0,
+            'conditional_logic' => 0,
+            'wrapper' => array(
+                'width' => '',
+                'class' => '',
+                'id' => '',
+            ),
+            'message' => __('Yes', 'municipio-customisation'),
+            'default_value' => 0,
+            'ui' => 0,
+            'ui_on_text' => '',
+            'ui_off_text' => '',
+        ),
     ),
     'location' => array(
         0 => array(