+++ views/modules/manualinput/base.blade.php
@@ -3,7 +3,7 @@
 @else
     @includeWhen(empty($hideTitle) && !empty($postTitle), 'partials.post-title')
     @if (!empty($manualInputs))
-        <div class="o-grid{{ !empty($stretch) ? ' o-grid--stretch' : '' }}">
+        <div class="o-grid{{ !empty($stretch) ? ' o-grid--stretch' : '' }}{{ !empty($centerContent) ? ' has-centered-content' : '' }}">
             @foreach ($manualInputs as $input)
                 @include('appearances.' . $input['view'])
             @endforeach